requirements.txt
langgraph>=0.2.50
langchain-core>=0.3.20
langchain-google-genai>=2.0.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.71.0
google-adk[a2a]>=0.3.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
rich>=13.7.0
